+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added Erlang distribution over serial (uart)
 
+### Fixed
+- Stop using deprecated `term_from_int32` on STM32 platform
+- Stop using deprecated `term_from_int32` on RP2 platform
+- Stop using deprecated `term_from_int32` on ESP32 platform
+
 ## [0.7.0-alpha.1] - 2026-04-06
 
 ### Added